pyproject.toml
[tool.poetry]
name = "earth-data-kit"
version = "0.1.3.dev20251101"
description = "Making GIS data analysis cheap and easy"
license = "Apache-2.0"
authors = ["Siddhant Gupta <siddhantgupta3@gmail.com>"]
readme = "README.md"

[tool.poetry_bumpversion.file."earth_data_kit/__init__.py"]
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[tool.poetry_bumpversion.file."docs/source/conf.py"]
search = 'release = "{current_version}"'
replace = 'release = "{new_version}"'

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.poetry.requires-plugins]
poetry-bumpversion = "^0.3.3"
poetry-plugin-export = ">=1.8"

[tool.poetry.dependencies]
python = ">=3.12,<4"
geopandas = "~0.14.4"
python-dotenv = "~1.0.1"
zarr = "~2.18.2"
netcdf4 = "~1.7.1.post2"
earthengine-api = "~1.5.2"
python-levenshtein = "~0.26.1"
lxml = "~5.3.1"
folium = "~0.19.5"
matplotlib = "~3.10.1"
xarray = "~2025.1.2"
dask = "~2025.7.0"
tqdm = "~4.67.1"
rio-tiler = "~7.6.0"
tenacity = "~9.1.2"
gdal = "==3.10.3"
pystac_client = "~0.8.6"
scipy = "^1.16.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "~6.29.4"
pytest = "~8.3.2"
sphinx = "~7.4.0"
furo = "~2024.5.6"
pytest-xdist = "~3.6.1"
black = "~25.1.0"
sphinx-copybutton = "~0.5.2"
pyarrow = "~19.0.1"
ipycytoscape = "~1.3.3"
gcsfs = "~2025.3.0"
bokeh = "~3.7.0"
pytest-order = "~1.3.0"
sphinxcontrib-googleanalytics = "^0.5"

[tool.poetry.plugins."xarray.backends"]
edk_dataset = "earth_data_kit.xarray_boosted.entrypoint:EDKDatasetBackend"